+++ Utilities.php
@@ -1,156 +1,180 @@
 <?php
-namespace DirectoyListing;
+namespace DirectoryListing;
+
+use Grav\Common\Grav;
+use Grav\Common\Plugin;
+use Grav\Common\Page\Page;
+use Grav\Common\Page\Media;
+use Grav\Common\Page\Collection;
 
 class Utilities
 {
-
     /**
-     * Unique identifier for HTML-structure
-     * @var int
+     * Plugin configuration
+     * @var array
      */
-    protected $seed;
+    protected $config;
 
     /**
      * Instantiate DirectoyListing Utilities
      */
-    public function __construct()
+    public function __construct($config)
     {
-        $this->seed = rand(1000, 9999);
+        $this->config = $config;
     }
 
     /**
-     * Builds hierarchical HTML list of files/folders
-     * @return string Hierarchical HTML List
+     * Sorts out configuration and initiates building
+     * @param string $route Route to page
+     * @return string HTML-list from page-structure
      */
-    public function buildDirectoryList($config, $page, $pages = false)
+    public function build($route)
     {
-        $css = $config['builtin_css'];
-        $javascript = $config['builtin_js'];
-        $links = $config['links'];
-        if (isset($config) && $config['enabled']) {
+        $config = $this->config;
+        $tree = $this->buildTree($route);
+        if (count(array_filter($config['include_additional'], 'strlen')) > 0) {
+            foreach ($config['include_additional'] as $include) {
+                $include = $this->buildTree($include, '@page.self');
+                $tree = array_merge($tree, $include);
+            }
+        }
+        if (count(array_filter($config['exclude_additional'], 'strlen')) > 0) {
+            foreach ($config['exclude_additional'] as $exclude) {
+                $this->arrayExcept($tree, array($exclude));
+            }
+        }
+        if ($tree) {
+            $list = $this->buildList($tree);
+            return $list;
+        } else {
+            return false;
+        }
+    }
+
+    /**
+     * Creates page-structure recursively
+     * @param string $route Route to page
+     * @param integer $depth Reserved placeholder for recursion depth
+     * @return array Page-structure with children and media
+     */
+    public function buildTree($route, $mode = false, $depth = 0)
+    {
+        $config = $this->config;
+        $page = Grav::instance()['page'];
+        $depth++;
+        $mode = '@page.self';
+        if ($depth > 1) {
+            $mode = '@page.children';
+        }
+        $pages = $page->evaluate([$mode => $route]);
+        $pages = $pages->published()->order($config['order']['by'], $config['order']['dir']);
+        $paths = array();
+        foreach ($pages as $page) {
+            if ($config['exclude_modular'] && isset($page->header()->content['items'])) {
+                if ($page->header()->content['items'] == '@self.modular') {
+                    continue;
+                }
+            }
+            $route = $page->rawRoute();
             $path = $page->path();
-            $title = $page->header()->title;
-            $excludeFiles = array();
-            if ($config['exclude_main']) {
-                $excludeFiles[] = $page->name();
-            }
-            if ($config['exclude_modular']) {
-                if (isset($page->header()->content['items'])) {
-                    if ($page->header()->content['items'] == '@self.modular') {
-                        $children = array_filter(array_keys($this->dirToArray($path)));
-                        foreach ($children as $module) {
-                            $excludeFiles[] = $module;
-                        }
-                    }
+            $title = $page->title();
+            $paths[$route]['depth'] = $depth;
+            $paths[$route]['title'] = $title;
+            $paths[$route]['route'] = $route;
+            $paths[$route]['name'] = $page->name();
+            if (!empty($paths[$route])) {
+                $children = $this->buildTree($route, $mode, $depth);
+                if (!empty($children)) {
+                    $paths[$route]['children'] = $children;
                 }
             }
-            if ($config['exclude_additional']) {
-                foreach ($config['exclude_additional'] as $fileToExclude) {
-                    $excludeFiles[] = $fileToExclude;
-                }
+            $media = new Media($path);
+            foreach ($media->all() as $filename => $file) {
+                $paths[$route]['media'][$filename] = $file->items()['type'];
             }
-            $output = '';
-            $items = array($title => $this->dirToArray($path, $excludeFiles));
-
-            if ($config['include_additional']) {
-                $instances = $pages;
-                foreach ($config['include_additional'] as $fileToInclude) {
-                    $includePath = $config['pages_path'] . '/' . $fileToInclude;
-                    if (in_array($includePath, array_keys($instances))) {
-                        $title = $instances[$includePath]->header()->title;
-                        $include = $this->dirToArray($includePath);
-                        $items[$title] = $include;
-                    }
-                }
-            }
-            foreach ($items as $key => $item) {
-                ob_start();
-                $seed = rand(1000, 9999);
-                echo '<div class="directorylist" id="directorylist_' . $seed . '">';
-                if (count($items) > 1) {
-                    echo '<ul class="directorylisting">';
-                    echo '<li class="item directory">';
-                    if ($css) {
-                        echo '<input type="checkbox" id="' . $seed . '_' . strtolower($key) . '" />';
-                        echo '<label for="' . $seed . '_' . strtolower($key) . '">' . $key . '</label>';
-                    } else {
-                        echo $key;
-                    }
-                    echo '</li>';
-                }
-                $this->recursiveArrayToList($page->rawRoute(), $item, $path, strtolower($key), $links, $css);
-                if (count($items) > 1) {
-                    echo '</ul>';
-                }
-                echo '</div>';
-                $output .= ob_get_contents();
-                ob_end_clean();
-            }
-
-            return $output;
+        }
+        if (!empty($paths)) {
+            return $paths;
+        } else {
+            return null;
         }
     }
 
     /**
-     * Directory-structure to array, recursively
-     * @param string $dir Path to folder
-     * @param array $excludes List of files/folders to exclude
-     * @return array Directory-structure
-     * @see http://php.net/manual/en/function.scandir.php#110570
+     * Creates HTML-lists recursively
+     * @param array $tree Page-structure with children and media
+     * @param integer $depth Reserved placeholder for recursion depth
+     * @return string HTML-list
      */
-    public function dirToArray($dir, $exclude = array())
+    public function buildList($tree, $depth = 0)
     {
-        $result = array();
-        $contents = array_filter(scandir($dir), function ($item) {
-            return $item[0] !== '.';
-        });
-        foreach ($contents as $value) {
-            if (!in_array($value, $exclude)) {
-                if (is_dir($dir . '/' . $value)) {
-                    $result[$value] = $this->dirToArray($dir . '/' . $value, $exclude);
-                } else {
-                    $result[] = $dir . '/' . $value;
-                }
-            }
+        $config = $this->config;
+        $depth++;
+        if ($config['builtin_css'] && $config['builtin_js'] && $depth == 1) {
+            $list = '<ul class="metismenu metisFolder">';
+        } else {
+            $list = '<ul>';
         }
-        return $result;
-    }
-
-    /**
-     * Directory-structure to HTML list
-     * @param array $array Directory-structure
-     * @param string $root Root path to files
-     * @param boolean $links Include links to files
-     * @param boolean $css Use plugin's styles
-     * @return void
-     * @see http://stackoverflow.com/a/36334789/603387
-     */
-    public function recursiveArrayToList($url, $array, $root, $uid = '', $links = false, $css = false)
-    {
-        echo '<ul class="directorylisting">';
-        foreach ($array as $key => $value) {
-            $location = str_replace($root, '', $value);
-            if (is_array($value)) {
-                if (!empty($value)) {
-                    echo '<li class="item directory">';
-                    if ($css) {
-                        echo '<input type="checkbox" id="' . $uid . '_' . $key . '" />';
-                        echo '<label for="' . $uid . '_' . $key . '">' . $key . '</label>';
-                    } else {
-                        echo $key;
-                    }
-                    echo '</li>';
-                }
-                $this->recursiveArrayToList($url, $value, $root, $uid . '_' . $key, $links, $css);
+        foreach ($tree as $route => $page) {
+            $list .= '<li class="directory';
+            if ($page['depth'] <= $config['level']) {
+                $list .= ' active';
+            }
+            $list .= '">';
+            if ($config['builtin_css'] && $config['builtin_js']) {
+                $list .= '<a href="#" aria-expanded="true" class="has-arrow">' . $page['title'] . '</a>';
             } else {
-                $parts = pathinfo($value);
-                if ($links) {
-                    echo '<li class="item file"><a href="' . $url . $location . '">' . $parts['basename'] . '</a></li>';
+                if ($config['links']) {
+                    $list .= '<a href="' . $page['route'] . '">' . $page['title'] . '</a>';
                 } else {
-                    echo '<li class="item file">' . $parts['basename'] . '</li>';
+                    $list .= $page['title'];
+                }
+            }
+            if (!$config['exclude_main']) {
+                if ($config['links']) {
+                    $list .= '<ul><li class="file page"><a href="' . $page['route'] . '">' . $page['name'] . '</a></li></ul>';
+                } else {
+                    $list .= '<ul><li class="file page">' . $page['name'] . '</li></ul>';
+                }
+            }
+            if (isset($page['children'])) {
+                $list .= $this->buildList($page['children'], $depth);
+            }
+            if (isset($page['media'])) {
+                $list .= '<ul>';
+                foreach ($page['media'] as $filename => $type) {
+                    if ($config['links']) {
+                        $list .= '<li class="file ' . $type . '"><a href="' . $page['route'] . '/' . $filename . '">' . $filename . '</a></li>';
+                    } else {
+                        $list .= '<li class="file ' . $type . '">' . $filename . '</li>';
+                    }
+                }
+                $list .= '</ul>';
+            }
+            $list .= '</li>';
+        }
+        $list .= '</ul>';
+        return $list;
+    }
+
+    /**
+     * Removes keys from an array
+     * @param array $array Array to remove keys from
+     * @param array $except Keys to remove
+     * @return void
+     * @see http://stackoverflow.com/a/37082098/603387
+     */
+    public function arrayExcept(&$array, $except)
+    {
+        foreach ($array as $key => $value) {
+            if (in_array($key, $except, true)) {
+                unset($array[$key]);
+            } else {
+                if (is_array($value)) {
+                    $this->arrayExcept($array[$key], $except);
                 }
             }
         }
-        echo '</ul>';
+        return;
     }
 }